fix BW reporting with mwait completion mode (w/o -j)
when running without the -j flag, the completion timestamp is used only when the
retry count is non-zero, a recent change 59c8acefa6d1, removed the retry
increment when using mwait in the completion check.

This patch fixes the regression by placing the monitor and retry counts in a
union.

Fixes: 59c8acefa6d1 ("fix cpu utilization for -x100")
Reported-by: Nirav N Shah <nirav.n.shah@intel.com>
Reviewed-by: Nirav N Shah <nirav.n.shah@intel.com>
Signed-off-by: Nikhil Rao <nikhil.rao@intel.com>