fix -j for descriptor count greater than queue depth
The commit below introduced a regression that causes a timeout when
the number of descriptors (batch descriptors when in batch mode)
exceeds the queue depth. In this case, the next descriptor to be
submitted needs to be saved at the end, and this code had been deleted
as part of the commit below.

This patch uses a different approach. In do_single_iter(),
descs[0..qd - 1/nb_desc - 1] remain in the WQ, when the function exits,
we use this invariant to simplify code (further details in code comments).

Reviewed-by: Nirav N Shah <nirav.n.shah@intel.com>
Fixes: 01fb7ec963db ("simplify code for -j (continous submission mode)")
Signed-off-by: Nikhil Rao <nikhil.rao@intel.com>